requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.2
Django==5.0.1
django-extensions==3.2.3
django-import-export==3.3.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn==21.2.0